WIP: shared folders mock
