+++ src/models/files/file-folder.js
@@ -1,65 +1,10 @@
-const { observable, computed } = require('mobx');
 const createMap = require('../../helpers/dynamic-array-map');
 const warnings = require('../warnings');
-const util = require('../../util');
+const AbstractFolder = require('./abstract-folder');
 
-class FileFolder {
-    @observable.shallow files = [];
-    @observable.shallow folders = [];
-    @observable name;
-    @observable createdAt;
-    @observable isDeleted;
-
-    // Variables for bulk actions and share-related actions
-    @observable selected;
-    @observable shareProgress;
-
-    get virtualFolders() {
-        return this.folders;
-    }
-
-    @computed get normalizedName() {
-        return this.name ? this.name.toLowerCase() : '';
-    }
-
-    @computed get foldersSortedByName() {
-        return this.virtualFolders.sort((f1, f2) => f1.normalizedName > f2.normalizedName);
-    }
-
-    @computed get filesSortedByDate() {
-        return this.files.sort((f1, f2) => f2.uploadedAt - f1.uploadedAt);
-    }
-
-    @computed get foldersAndFilesDefaultSorting() {
-        const { foldersSortedByName, filesSortedByDate } = this;
-        return foldersSortedByName.concat(filesSortedByDate);
-    }
-
-    @computed get size() {
-        let currentSize = 0;
-        for (const folder of this.folders) {
-            currentSize += folder.size;
-        }
-        for (const file of this.files) {
-            currentSize += file.size;
-        }
-        return currentSize;
-    }
-
-    @computed get sizeFormatted() {
-        return util.formatBytes(this.size);
-    }
-
-    @observable parent = null;
-    isFolder = true;
-    get isRoot() {
-        return !this.parent;
-    }
-    get hasNested() {
-        return this.virtualFolders && this.virtualFolders.length;
-    }
-
+class FileFolder extends AbstractFolder {
     constructor(name) {
+        super();
         const m = createMap(this.files, 'fileId');
         this.name = name;
         this.fileMap = m.map;
